Factor the CLI helper functions into their own files
